feat: adding more functions.

+++ desafio-sinapses-kai/desafio/projeto base/sinapses/src/main/java/br/com/zaffari/sinapses/repository/SinapseRepository.java
@@ -8,6 +8,6 @@ import br.com.zaffari.sinapses.domain.Sinapse;
 import java.util.List;
 
 @Repository
-public interface SinapseRepository extends JpaRepository<Sinapse, Integer> {
-    List<Sinapse> findSinapseById(Integer id);
+public interface SinapseRepository extends JpaRepository<Sinapse, Long> {
+    List<Sinapse> findSinapseById(Long id);
 }
